+++ NEWS.md
@@ -1,6 +1,9 @@
 
 # development version
 
+* cleancall now better supports embedding it into another R package, see
+  the 'Embedding cleancall' section in the README for details.
+
 # 0.1.1
 
 * Fix compilation errors with `-fno-common`.